Update: Refactoring
